fixed new user flow
