+++ package.json
@@ -11,11 +11,11 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.10.0",
-    "@prisma/client": "^6.14.0",
+    "@prisma/client": "^6.16.2",
     "@prisma/extension-accelerate": "^2.0.2",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-dropdown-menu": "^2.1.15",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-popover": "^1.1.15",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tooltip": "^1.2.8",
@@ -23,30 +23,30 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "date-fns": "^4.1.0",
-    "lucide-react": "^0.539.0",
-    "next": "15.4.6",
+    "lucide-react": "^0.544.0",
+    "next": "15.5.3",
     "next-auth": "^5.0.0-beta.29",
     "next-themes": "^0.4.6",
-    "prisma": "^6.14.0",
+    "prisma": "^6.16.2",
     "qrcode.react": "^4.2.0",
-    "react": "19.1.0",
+    "react": "19.1.1",
     "react-confetti": "^6.4.0",
-    "react-day-picker": "^9.8.1",
-    "react-dom": "19.1.0",
+    "react-day-picker": "^9.10.0",
+    "react-dom": "19.1.1",
     "react-use": "^17.6.0",
-    "recharts": "^3.1.2",
+    "recharts": "^3.2.1",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "15.4.6",
-    "tailwindcss": "^4",
-    "tw-animate-css": "^1.3.6",
-    "typescript": "^5"
+    "@eslint/eslintrc": "^3.3.1",
+    "@tailwindcss/postcss": "^4.1.13",
+    "@types/node": "^24.5.0",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
+    "eslint": "^9.35.0",
+    "eslint-config-next": "15.5.3",
+    "tailwindcss": "^4.1.13",
+    "tw-animate-css": "^1.3.8",
+    "typescript": "^5.9.2"
   }
 }
